Initial naive implementation to pass test 2
The naive implementation is to just call the royalty department with the
packing slip. The problem with this is that they are only interested
when the product is a book.

Additionally for some reason I made the Product on the order a
collection (not sure why). This will be resolved in following commits.